pyproject.toml
[tool.poetry]
name = "treinamento-simcc"
version = "0.1.0"
description = ""
authors = ["meirelescg <geu_costa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
